+++ src/ocaml/typing/printtyped.ml
@@ -975,3 +975,5 @@ let implementation ppf x = list 0 structure_item ppf x.str_items;;
 
 let implementation_with_coercion ppf Typedtree.{structure; _} =
   implementation ppf structure
+
+let expression ppf x = expression 0 ppf x